// packages/shared-types/src/enrichissement/enums/code-erreur.enum.ts
export enum CodeErreurEnrichissement {
  CADASTRE_INTROUVABLE = "CADASTRE_INTROUVABLE",
  CADASTRE_API_ERROR = "CADASTRE_API_ERROR",
  BDNB_API_ERROR = "BDNB_API_ERROR",
  ENEDIS_API_ERROR = "ENEDIS_API_ERROR",
  TRANSPORT_API_ERROR = "TRANSPORT_API_ERROR",
  OVERPASS_API_ERROR = "OVERPASS_API_ERROR",
  LOVAC_API_ERROR = "LOVAC_API_ERROR",
  ENRICHISSEMENT_FAILED = "ENRICHISSEMENT_FAILED",
  UNKNOWN_ERROR = "UNKNOWN_ERROR",
  GEORISQUES_RGA_ERROR = "GEORISQUES_RGA_ERROR",
}